mcp-server/requirements.txt
fastmcp
uvicorn
google-cloud-aiplatform

